pyproject.toml
[project]
name = "systems-thinking-plugin"
version = "0.1.0"
description = "Claude Code plugin for infrastructure and network architecture decision workflows"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
test = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"utils/*.py" = ["E501"]  # allow long lines in docstring usage examples

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "eval: marks tests as eval tests (may require claude CLI)",
    "slow: marks tests as slow-running",
]
addopts = "-v --tb=short"
